Downgrade Go version requirement to 1.20
As part of the upgrade to Go 1.21, the Go toolchain now requires the
`go` directive to match the maximum Go version in use in dependencies.

This leads to any transitive dependency on this library to result in a
requirement of the consuming project moving to Go 1.21.

Although in a lot of cases this may not be as problematic, it forces
consumers to migrate, which we don't need to do in this case.